Update site-recovery-vmware-to-azure.md
Went through an issue with a customer right now where the R3 version of PowerCLI broke the configuration of the server.  We had to rebuild the processing server completely as when you first install the R3 version you cannot uninstall and revert to version 6.0, you have to rebuild the whole server which wasted significant time on the project trying to troubleshoot.